@@ -1,4 +1,4 @@
-class CreateRecipes < ActiveRecord::Migration
+class CreateRecipes < ActiveRecord::Migration[5.0]
   def change
     create_table :recipes do |t|
       t.string :title
