@@ -1,4 +1,4 @@
-class RenameUserPasswordToPasswordDigest < ActiveRecord::Migration
+class RenameUserPasswordToPasswordDigest < ActiveRecord::Migration[5.0]
   def change
     rename_column :users, :password, :password_digest
   end
