@@ -1,4 +1,4 @@
-class MakeUserEmailUnique < ActiveRecord::Migration
+class MakeUserEmailUnique < ActiveRecord::Migration[5.0]
   def change
     add_index :users, :email, unique: true
   end
