@@ -1,4 +1,4 @@
-class AddUserIdToRecipe < ActiveRecord::Migration
+class AddUserIdToRecipe < ActiveRecord::Migration[5.0]
   def change
     add_column :recipes, :user_id, :integer
   end
