@@ -1,4 +1,4 @@
-class AddAdminToUser < ActiveRecord::Migration
+class AddAdminToUser < ActiveRecord::Migration[5.0]
   def change
     add_column :users, :admin, :integer, default: 0
   end
